requirements.txt
scrapy>=2.11.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0
tqdm>=4.64.0
playwright>=1.40.0
pyyaml>=6.0.1
jsonschema>=4.18.0
jinja2>=3.1.2
python-dateutil>=2.8.2